Transferred the Networking codes to APIService.
